fixes filesystem spec, reorganizes code
