+++ sources/standard/filesystem/settings.py
@@ -1,1 +1,1 @@
-DEFAULT_CHUNK_SIZE = 10
+DEFAULT_CHUNK_SIZE = 100
