+++ sources/standard/requirements.txt
@@ -1,2 +1,1 @@
 dlt>=0.3.18,<0.4
-adlfs>=2023.9.0
